refactor: AuthenticationPrincipal 및 관련 클래스 config.auth 패키지로 이동
AuthenticationPrincipal, ArgumentResolver, Interceptor 등의 auth 관련 클래스를 config.auth 패키지로 이동하여 패키지 구조를 개선했습니다. 경로 변경에 따른 import 구문도 수정되었습니다.
