+++ src/main/java/roomescape/auth/AuthController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.dto.LoginMember;
 import roomescape.auth.dto.LoginRequest;
 import roomescape.auth.dto.LoginResponse;
+import roomescape.config.auth.AuthenticationPrincipal;
 
 @RestController
 public class AuthController {
