+++ src/main/java/roomescape/config/auth/AuthorizationAdminInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.config.auth;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -7,6 +7,8 @@ import java.util.Arrays;
 import java.util.Objects;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.servlet.HandlerInterceptor;
+import roomescape.auth.JwtProvider;
+import roomescape.auth.TokenBody;
 import roomescape.exception.custom.reason.auth.AuthNotExistsCookieException;
 import roomescape.exception.custom.reason.auth.AuthNotValidTokenException;
 import roomescape.member.MemberRole;
